avant merge

+++ appsae/gestion_note.py
@@ -1,9 +1,31 @@
 from .models import *
 from django.db.models import Avg
+from django.conf import settings
+from csv import writer
+
+
+def addavisCSV(avis):
+    list = [str(avis.adherant_fk_id), ' ' + str(avis.restaurant_fk_id), ' ' + str(float(avis.note))]
+
+    with open(str(settings.BASE_DIR) + '/' + "ratings.csv", 'a') as f_object:
+        writer_object = writer(f_object)
+        writer_object.writerow(list)
+        f_object.close()
+
+
+def avisExist(user, restaurant):
+    """ Vérifie si l'utilisateur user a déjà ajouté un avis sur le restaurant
+
+    @param user: l'utilisateur
+    @param restaurant: le restaurant
+    @return: true si l'avis existe, false sinon
+    """
+    return Avis.objects.filter(restaurant_fk=restaurant, adherant_fk=user).count() != 0
 
 
 def updateNoteMoyenneRestaurant(restaurant):
     """ Fonction de mise à jour de la note moyenne d'un restaurant passé en paramètres
+
     @param nomRestaurant: le nom du restaurant
     @return: /
     """
@@ -21,9 +43,16 @@ def ajoutAvis(user, restaurant, note, avis):
     @param restaurant: le restaurant
     @return: /
     """
-    ajout = Avis(note=note, texte=avis, restaurant_fk=restaurant, adherant_fk=user)
-    ajout.save()
-    updateNoteMoyenneRestaurant(restaurant)
+    if avisExist(user, restaurant) == False:
+        nb_review_ad = Adherant.objects.get(pk=user.pk).nb_review
+        print("nb review adherannt :" + str(nb_review_ad))
+        ajout = Avis(note=note, texte=avis, restaurant_fk=restaurant, adherant_fk=user)
+        ajout.save()
+        updateNoteMoyenneRestaurant(restaurant)
+        if nb_review_ad > 5: # & tps dans la bd < à 200 secondes
+            # algorecommendation
+            a = 1
+        addavisCSV(ajout)
 
 
 def updateAvis(user, restaurant, note, avis):
@@ -34,8 +63,11 @@ def updateAvis(user, restaurant, note, avis):
     @param avis: le nouvel avis
     @return: /
     """
-    Avis.objects.filter(restaurant_fk=restaurant, adherant_fk=user).update(note=note, texte=avis)
-    updateNoteMoyenneRestaurant(restaurant)
+    if avisExist(user,restaurant):
+        Avis.objects.filter(restaurant_fk=restaurant, adherant_fk=user).update(note=note, texte=avis)
+        updateNoteMoyenneRestaurant(restaurant)
+
+
 def suppressionAvis(user, restaurant):
     """ Suppression de l'avis de l'utilisateur user sur le restaurant restaurant
 
@@ -43,33 +75,34 @@ def suppressionAvis(user, restaurant):
     @param restaurant: le restaurant
     @return: /
     """
-    Avis.objects.get(restaurant_fk=restaurant, adherant_fk=user).delete()
-    updateNoteMoyenneRestaurant(restaurant)
+    if avisExist(user,restaurant):
+        Avis.objects.get(restaurant_fk=restaurant, adherant_fk=user).delete()
+        updateNoteMoyenneRestaurant(restaurant)
 
-def supplettreUTF():
+
+def listeAffichageAvis(restaurant, num, user=""):
+    """ Renvoie une liste d'avis 10 par 10 ne contenant pas l'avis de l'utilisateur user,
+    si num vaut 0, on renvoie de 0 à 9 dans la liste des avis, etc...
+
+    @param restaurant: le restaurant
+    @param user: l'utilisateur
+    @param num: le numéro de la page
+    @return: une liste (QuerySet) d'avis
     """
+    taille_list = 2
+    if user == "":
+        avis = Avis.objects.filter(restaurant_fk=restaurant)
+    else:
+        avis = Avis.objects.filter(restaurant_fk=restaurant).exclude(adherant_fk=user)
+    return avis[num*taille_list:(num + 1)*taille_list]
 
-    @return:
+
+def afficherVoirPlus(restaurant, num, user=""):
+    """Renvoie true s'il faut afficher le bouton "Voir Plus", false sinon
+
+    @param restaurant: le restaurant concerné
+    @param num: le numéro de la page actuelle
+    @param user: l'utilisateur concerné
+    @return: booléen en fonction de s'il faut afficher ou non "Voir Plus"
     """
-    for resto in Restaurant.objects.all():
-        nouveau_nom = testNomUTF(resto.nom)
-        if (nouveau_nom != resto.nom):
-            Restaurant.objects.filter(id_yelp=resto.id_yelp).update(nom=nouveau_nom)
-
-
-def testNomUTF(nom):
-    """
-
-    @param nom:
-    @return:
-    """
-    list = "azertyuiopqsdfghjklmwxcvbnAZERTYUIOPQSDFGHJKLMWXCVBN0123456789&'-_+/:,*²#|!?°. "
-    nouveau_nom = ""
-    for lettre in nom:
-        if lettre in list:
-            nouveau_nom += lettre
-        elif lettre in 'éèê':
-            nouveau_nom += 'e'
-        elif lettre in 'ÉÈ':
-            nouveau_nom += 'E'
-    return nouveau_nom
+    return listeAffichageAvis(restaurant, num + 1, user).count() != 0
